Support showing previews in the photo viewer.
The photo viewer now supports the notion of a
preview of the full image. The preview is not
pinchable in any manner and displays a circular
progress bar over top of it while we are downloading
the full version of the image.

To enable this functionality, there is a new column
in the PhotoContract API called thumbnailUri. If it is
not set to null and if the attempted load of the contentUri
is unsuccessful, we will attempt to load a thumbnail version
that will show until the contentUri is finished loading.

Change-Id: I5547a5fc6a30ee6a30f137dc1dee76a77fb5b304