Add HTTP uri handling & fetch thumbnail first
If a http:// or https:// url is specified as the thumbnail uri or
regular uri for an image, we attempt to fetch it over the internet.

Also, changed the order of loading so that we always attempt to fetch
the thumbnail image before we fetch the regular image. This aids in the
http:// uri case, since it should take significantly less time to be
presented with a thumbnail than the full image.

Finally, finishes the photo viewer activity when the "top" button on the
action bar is pressed, unless a custom override has been added.

These 3 changes could be seperated into different patches if you would
like

Change-Id: Iddc8cc6bef3f9e66198dfb98eb2f7bd309023b3f