Add option to permenantly show ProgressBar
In order to indicate to the user that more content is loading, the
provider can provide an image with the loading field set and null as it's
url. In this case, we will not try loading any image, and will just show the
ProgressBar.

Something like this is nessecary in Velvet, where we wish to load
infinite images.

Change-Id: Ib26412a579abdbe7fbd338b99ff39f4d1c13f806