Fix scale popping by disabling zoom on thumbnails
In order to prevent the situation where a user scales a low-res image in
the image viewer before the new image has loaded, we disable the scaling
and panning of low-res images.

This is desirable because when the high-res image loads, it will "pop"
back to it's default scale and pan settings.

In the longer term I would like to fix this problem by making high-res
images load in at the current scale and pan settings. This is a much
quicker fix, however.

Bug: 7338705

Change-Id: I0c8f3dc5a7b29943db9c03e9c82cc3cee5ce2734