Performance improvement for photoViewer
I noticed that when we bring up the photoViewer,
it always initializes the first two fragments, and
then switches to whatever fragments we actually
want to look at based upon the intent parameters.
This issue is that we were assigning the adapter
to the ViewPager before we had any cursor at all.
If we hold off on setting the adpater until we have
a cursor, then this doesn't happen.  This improves
startup time of the PhotoViewActivity noticeably.

Also, made a few changes to make it easier to
extend PhotoViewActivity and PhotoViewFragment,
to change its behavior.

Change-Id: Icc4c397529bd147d0ac6818f79a48db1c4a25f32