Save/Restore Actionbar title
In rare cases, we cannot restore the actionbar title in the photo
viewer because we require the fragment to be visible on the screen to
update the title.

Since this is a rather rare problem, the easiest fix is to save the
actionbar title to a bundle and restore it on orientation changes.

Bug: 8124251 Action bar text for the attachment disappears on
             orientation change

Change-Id: I27b2768b3e7edb18e1ddd272f48a3ad8ebdd408a