Use ConnectionManager to restart loaders
When we come back from having no connection, but we had failed to load
anything before, we don't bother trying to restart our loaders.

This change ensures that if we get notified that the device now has an
active internet connection, and we haven't bound a photo (or a
thumbnail), then we restart the appropriate loaders and show the loading
indicator.

Bug: 8407579

Change-Id: I30295ca9585839760421fb37c3a28c987357aacd