Update PhotoViewer so that Babel can use it
The photoViewer now supports the needs of the Babel app.
In particular:
1. It allows for applications to extend the PhotoBitmapLoader
for their own needs. Babel needs this because it already
maintains a customized image cache in memory, and we don't
want to have duplicate bitmaps in memory.
2. The photoViewActivity can optionally use an animation
that scales the image up from the image's original location
on screen to the full size.
3. The photoViewActivity can optionally be in fullscreen
mode by default, rather than switching to fullscreen
when tapped.

Change-Id: Id5d62f1961c1ce2b9402c7fb19583baad4d96580