Bug 18864683: Accessibility
The actionbar is made to stay visibile if accessibility is enabled. This
allows the user to explore and access the menu buttons without it auto-hiding.

The photo details are announced whenever switching to a new photo.

The root view of photo_activity_view is removed from touch exploration as
no actions can be performed by touching it.

Change-Id: I9090920e2154cd9f038d2bede0e4ec9399becc3d