Fixed photo unbinding issue in PhotoViewFragment class
When the fragment becomes inactive, for example due to a system dialog
popup, the image that was shown on the fragment disappears.

The reason for this behavior is that photo is unbinded in #onPause.

To fix this, photo unbinding operation is removed, because unbinding
is an unnecessary operation.

This change fixes the following bug: b/26436886

Change-Id: Ia5ca3f8ebafbd5de4828744a29fd878f32e9236d