PhotoViewer sets ContentDescriptions when provided by PagerAdapter
Right now, when the user is navigating using accessibility, then when
they open a tab it announces the image, which is awesome.
If they then navigate to each view, there's two unlabeled views, so
accessibility users don't know what they have selected, which is
a GAR violation. One of which is the PhotoViewFragment's view, which
accepts touch (for zooming and sliding), so we can't tell
accessibility to ignore it.
(https://screenshot.googleplex.com/9p3jFT39QV9.png)
This change adds a contentDescription to the image view, so visually
impaired users can tell which view currently has focus.

This does not fix the other unlabeled view which I haven't found yet

Bug: 27686286
Change-Id: Ie9acd5100cb19784523ccee4d0af3bf311c3231f